extract world controls component
