+++ src/ui/views/session/world/WorldView.tsx
@@ -1,70 +1,35 @@
-import { useCallback, useEffect, useRef, useState, forwardRef } from "react";
-import { GroupCall, Room, RoomStatus } from "@thirdroom/hydrogen-view-sdk";
+import { useEffect, useState } from "react";
+import { Room } from "@thirdroom/hydrogen-view-sdk";
 import classNames from "classnames";
-import { useAtom, useAtomValue, useSetAtom } from "jotai";
+import { useAtom, useAtomValue } from "jotai";
 
 import { WorldChat } from "../world-chat/WorldChat";
 import { Stats } from "../stats/Stats";
 import { Text } from "../../../atoms/text/Text";
-import { IconButton } from "../../../atoms/button/IconButton";
 import { useKeyDown } from "../../../hooks/useKeyDown";
 import { useEvent } from "../../../hooks/useEvent";
-import PeopleIC from "../../../../../res/ic/peoples.svg";
-import SubtitlesIC from "../../../../../res/ic/subtitles.svg";
-import SubtitlesOffIC from "../../../../../res/ic/subtitles-off.svg";
-import MicIC from "../../../../../res/ic/mic.svg";
-import MicOffIC from "../../../../../res/ic/mic-off.svg";
-import CrossIC from "../../../../../res/ic/cross.svg";
-import CallCrossIC from "../../../../../res/ic/call-cross.svg";
-import HelpIC from "../../../../../res/ic/help.svg";
-import XRIC from "../../../../../res/ic/xr.svg";
 import "./WorldView.css";
 import { EditorView } from "../editor/EditorView";
-import { useCallMute } from "../../../hooks/useCallMute";
-import { Tooltip } from "../../../atoms/tooltip/Tooltip";
 import { Reticle } from "../reticle/Reticle";
-import { EntityTooltip, IPortalProcess } from "../entity-tooltip/EntityTooltip";
+import { EntityTooltip } from "../entity-tooltip/EntityTooltip";
 import { Nametags } from "../nametags/Nametags";
-import { Dialog } from "../../../atoms/dialog/Dialog";
-import { MemberListDialog } from "../dialogs/MemberListDialog";
 import { useMainThreadContext } from "../../../hooks/useMainThread";
-import { registerMessageHandler, Thread } from "../../../../engine/module/module.common";
-import { NametagsEnableMessage, NametagsEnableMessageType } from "../../../../plugins/nametags/nametags.common";
+import { registerMessageHandler } from "../../../../engine/module/module.common";
 import { useToast } from "../../../hooks/useToast";
-import { usePermissionState } from "../../../hooks/usePermissionState";
-import { useMicrophoneState } from "../../../hooks/useMicrophoneState";
 import { useHydrogen } from "../../../hooks/useHydrogen";
-import { exceptionToString, RequestException, useStreamRequest } from "../../../hooks/useStreamRequest";
-import { AlertDialog } from "../dialogs/AlertDialog";
-import { OnboardingModal } from "./OnboardingModal";
-import { useOnboarding } from "../../../hooks/useOnboarding";
-import { Header } from "../../../atoms/header/Header";
-import { HeaderTitle } from "../../../atoms/header/HeaderTitle";
-import { ShortcutUI } from "./ShortcutUI";
-import { Scroll } from "../../../atoms/scroll/Scroll";
 import { IMainThreadContext } from "../../../../engine/MainThread";
-import { useMouseDown } from "../../../hooks/useMouseDown";
-import {
-  InteractableAction,
-  InteractionMessage,
-  InteractionMessageType,
-} from "../../../../plugins/interaction/interaction.common";
-import { ExitedWorldMessage, ThirdRoomMessageType } from "../../../../plugins/thirdroom/thirdroom.common";
 import { createDisposables } from "../../../../engine/utils/createDisposables";
-import { aliasToRoomId, parsedMatrixUriToString, parseMatrixUri } from "../../../utils/matrixUtils";
-import { Hotbar, HotbarSlot } from "../../components/hotbar/Hotbar";
 import { ObjectCapReachedMessage, ObjectCapReachedMessageType } from "../../../../plugins/spawnables/spawnables.common";
-import { useWorldAction } from "../../../hooks/useWorldAction";
 import { useCalls } from "../../../hooks/useCalls";
 import { useRoomCall } from "../../../hooks/useRoomCall";
-import { useIsMounted } from "../../../hooks/useIsMounted";
 import { InteractableType } from "../../../../engine/resource/schema";
 import { useWebXRSession } from "../../../hooks/useWebXRSession";
-import { useMemoizedState } from "../../../hooks/useMemoizedState";
-import { overlayWorldAtom } from "../../../state/overlayWorld";
 import { worldChatVisibilityAtom } from "../../../state/worldChatVisibility";
 import { overlayVisibilityAtom } from "../../../state/overlayVisibility";
 import { worldAtom } from "../../../state/world";
+import { HotbarControls, WorldControls } from "./WorldControls";
+import { WorldOnboarding } from "./WorldOnboarding";
+import { useLocalStorage } from "../../../hooks/useLocalStorage";
 
 export interface ActiveEntityState {
   interactableType: InteractableType;
@@ -75,51 +40,6 @@ export interface ActiveEntityState {
 }
 
 const SHOW_NAMES_STORE = "showNames";
-
-const MuteButton = forwardRef<HTMLButtonElement, { activeCall?: GroupCall; showToast: (text: string) => void }>(
-  ({ activeCall, showToast }, ref) => {
-    const { platform } = useHydrogen(true);
-    const micPermission = usePermissionState("microphone");
-    const requestStream = useStreamRequest(platform, micPermission);
-    const [micException, setMicException] = useState<RequestException>();
-    const [microphone, setMicrophone] = useMicrophoneState();
-    const { mute: callMute, handleMute } = useCallMute(activeCall);
-    if (callMute === microphone) {
-      setMicrophone(!microphone);
-    }
-
-    return (
-      <>
-        {micException && (
-          <AlertDialog
-            open={!!micException}
-            title="Microphone"
-            content={<Text variant="b2">{exceptionToString(micException)}</Text>}
-            requestClose={() => setMicException(undefined)}
-          />
-        )}
-        <Tooltip content={callMute ? "Unmute" : "Mute"}>
-          <IconButton
-            variant="world"
-            label="Mic"
-            iconSrc={callMute ? MicOffIC : MicIC}
-            onClick={() => {
-              showToast(!callMute ? "Microphone Muted" : "Microphone Unmuted");
-              handleMute(async () => {
-                const [stream, exception] = await requestStream(true, false);
-                if (stream) return stream;
-                setMicException(exception);
-                return undefined;
-              });
-            }}
-            ref={ref}
-          />
-        </Tooltip>
-      </>
-    );
-  }
-);
-
 interface WorldViewProps {
   world: Room;
 }
@@ -129,235 +49,80 @@ export function WorldView({ world }: WorldViewProps) {
   const { session } = useHydrogen(true);
   const calls = useCalls(session);
   const activeCall = useRoomCall(calls, world.id);
-  const { enterWorld, exitWorld } = useWorldAction(session);
-  const selectWorld = useSetAtom(overlayWorldAtom);
   const isWorldEntered = useAtomValue(worldAtom).entered;
   const [worldChatVisible, setWorldChatVisibility] = useAtom(worldChatVisibilityAtom);
   const [overlayVisible, setOverlayVisibility] = useAtom(overlayVisibilityAtom);
   const [editorEnabled, setEditorEnabled] = useState(false);
   const [statsEnabled, setStatsEnabled] = useState(false);
-  const [shortcutUI, setShortcutUI] = useState(false);
-  const isMounted = useIsMounted();
 
-  const { onboarding, finishOnboarding } = useOnboarding(isWorldEntered ? world?.id : undefined);
-
-  const muteBtnRef = useRef<HTMLButtonElement | null>(null);
   const { toastShown, toastContent, showToast } = useToast();
 
   const engine = useMainThreadContext();
-
-  const [showNames, setShowNames] = useState<boolean>(() => {
-    const store = localStorage.getItem(SHOW_NAMES_STORE);
-    if (!store) return true;
-    const json = JSON.parse(store);
-    return json.showNames;
-  });
-
-  useEffect(() => {
-    localStorage.setItem(SHOW_NAMES_STORE, JSON.stringify({ showNames }));
-
-    engine.sendMessage<NametagsEnableMessageType>(Thread.Game, {
-      type: NametagsEnableMessage,
-      enabled: showNames,
-    });
-  }, [engine, showNames]);
-
-  const [showActiveMembers, setShowActiveMembers] = useState<boolean>(false);
-
-  const toggleShowNames = useCallback(() => {
-    const enabled = !showNames;
-    setShowNames(enabled);
-    engine.sendMessage<NametagsEnableMessageType>(Thread.Game, { type: NametagsEnableMessage, enabled });
-    showToast(enabled ? "Show Names" : "Hide Names");
-  }, [setShowNames, showNames, showToast, engine]);
-
-  const toggleShowActiveMembers = () => setShowActiveMembers((state) => !state);
-  const toggleShortcutUI = () => setShortcutUI((state) => !state);
-  const [activeEntity, setActiveEntity] = useMemoizedState<ActiveEntityState | undefined>();
-  const [portalProcess, setPortalProcess] = useState<IPortalProcess>({});
-  const mouseDown = useMouseDown(mainThread.canvas);
-
+  const [showNames, setShowNames] = useLocalStorage(SHOW_NAMES_STORE, true);
   const { isWebXRSupported, enterXR, isPresenting } = useWebXRSession();
 
   useEffect(() => {
-    let unSubStatusObserver: () => void | undefined;
-
-    const onInteraction = async (ctx: IMainThreadContext, message: InteractionMessage) => {
-      const interactableType = message.interactableType;
-
-      if (!interactableType || message.action === InteractableAction.Unfocus) {
-        setActiveEntity(undefined);
-      } else if (
-        message.interactableType === InteractableType.Grabbable ||
-        message.interactableType === InteractableType.Interactable ||
-        message.interactableType === InteractableType.UI
-      ) {
-        setActiveEntity({
-          interactableType,
-          name: message.name || "Object",
-          held: message.held || false,
-          ownerId: message.ownerId,
-        });
-      } else if (message.interactableType === InteractableType.Player) {
-        if (message.action === InteractableAction.Grab) {
-          setShowActiveMembers(true);
-          document.exitPointerLock();
-        } else {
-          setActiveEntity({
-            interactableType,
-            name: (message.peerId && activeCall?.members.get(message.peerId)?.member.displayName) || "Player",
-            peerId: message.peerId,
-            held: false,
-          });
-        }
-      } else if (message.interactableType === InteractableType.Portal) {
-        if (message.action === InteractableAction.Grab) {
-          try {
-            setPortalProcess({});
-            if (!message.uri) throw Error("Portal does not have valid matrix id/alias");
-
-            const parsedUri = parseMatrixUri(message.uri);
-            if (parsedUri instanceof URL) {
-              window.location.href = parsedUri.href;
-              return;
-            }
-
-            const roomIdOrAlias = parsedUri.mxid1;
-            const roomId = roomIdOrAlias.startsWith("#")
-              ? aliasToRoomId(session.rooms, parsedUri.mxid1)
-              : parsedUri.mxid1;
-
-            if (roomId && session.rooms.get(roomId)) {
-              selectWorld(roomId);
-              enterWorld(roomId);
-              return;
-            }
-
-            setPortalProcess({ joining: true });
-            const rId = await session.joinRoom(roomIdOrAlias);
-            if (!isMounted()) return;
-            setPortalProcess({});
-            const roomStatusObserver = await session.observeRoomStatus(rId);
-            unSubStatusObserver = roomStatusObserver.subscribe((roomStatus) => {
-              if (roomStatus !== RoomStatus.Joined) return;
-              selectWorld(rId);
-              enterWorld(rId);
-            });
-          } catch (err) {
-            if (!isMounted()) return;
-            setPortalProcess({ error: err as Error });
-          }
-        } else {
-          setActiveEntity({
-            interactableType,
-            name: (message.uri && parsedMatrixUriToString(parseMatrixUri(message.uri))) || "Portal",
-            held: false,
-          });
-        }
-      }
-    };
-
-    const onExitedWorld = (ctx: IMainThreadContext, message: ExitedWorldMessage) => {
-      setActiveEntity(undefined);
-    };
-
     const onObjectCapReached = (ctx: IMainThreadContext, message: ObjectCapReachedMessage) => {
       showToast("Maximum number of objects reached.");
     };
 
     const disposables = createDisposables([
-      registerMessageHandler(engine, InteractionMessageType, onInteraction),
       registerMessageHandler(engine, ObjectCapReachedMessageType, onObjectCapReached),
-      registerMessageHandler(engine, ThirdRoomMessageType.ExitedWorld, onExitedWorld),
     ]);
     return () => {
-      unSubStatusObserver?.();
       disposables();
     };
-  }, [activeCall, engine, showToast, enterWorld, selectWorld, isMounted, session, setActiveEntity]);
+  }, [engine, showToast]);
 
   useKeyDown(
     (e) => {
-      if (isWorldEntered === false) return;
-      if (onboarding) return;
+      const inputFocused = document.activeElement?.tagName.toLowerCase() === "input";
 
-      const isEscape = e.key === "Escape";
-      const isTyping = document.activeElement?.tagName.toLowerCase() === "input";
+      if (e.key === "Escape") {
+        if (worldChatVisible) {
+          mainThread.canvas?.requestPointerLock();
+          setWorldChatVisibility(false);
+          return;
+        }
 
-      if (isEscape && showActiveMembers) {
-        mainThread.canvas?.requestPointerLock();
-        setShowActiveMembers(false);
-        return;
+        if (inputFocused) return;
+
+        if (editorEnabled) {
+          mainThread.canvas?.requestPointerLock();
+          setEditorEnabled(false);
+          return;
+        }
+
+        if (overlayVisible) {
+          mainThread.canvas?.requestPointerLock();
+          setOverlayVisibility(false);
+          return;
+        } else {
+          document.exitPointerLock();
+          setOverlayVisibility(true);
+          return;
+        }
       }
-      if (isEscape && shortcutUI) {
-        mainThread.canvas?.requestPointerLock();
-        setShortcutUI(false);
-        return;
-      }
-      if (isEscape && worldChatVisible) {
-        mainThread.canvas?.requestPointerLock();
-        setWorldChatVisibility(false);
-        return;
-      }
-      if (isEscape && overlayVisible) {
-        mainThread.canvas?.requestPointerLock();
-        setOverlayVisibility(false);
-        return;
-      }
-      if (isEscape && overlayVisible === false) {
-        document.exitPointerLock();
-        setOverlayVisibility(true);
-        return;
-      }
-      if (e.key === "Enter" && overlayVisible === false && worldChatVisible === false) {
+
+      if (inputFocused) return;
+
+      if (e.key === "Enter" && !overlayVisible && !worldChatVisible) {
         if (document.activeElement !== document.body) return;
         document.exitPointerLock();
         setWorldChatVisibility(true);
         return;
       }
 
-      if (isTyping || worldChatVisible || showActiveMembers || shortcutUI) return;
-
-      if (e.altKey && e.code === "KeyL") {
-        exitWorld();
-      }
-      if (e.altKey && e.code === "KeyX" && isWebXRSupported) {
-        enterXR();
-      }
-      if (e.code === "KeyM" && muteBtnRef.current !== null) {
-        muteBtnRef.current.click();
-      }
       if (e.code === "Backquote") {
         setEditorEnabled((enabled) => !enabled);
+        return;
       }
       if (e.code === "KeyS" && e.shiftKey && e.ctrlKey) {
         setStatsEnabled((enabled) => !enabled);
-      }
-      if (e.code === "KeyN") {
-        toggleShowNames();
-      }
-      if (e.code === "KeyP") {
-        document.exitPointerLock();
-        toggleShowActiveMembers();
-      }
-      if (e.code === "Slash") {
-        e.preventDefault();
-        toggleShortcutUI();
+        return;
       }
     },
-    [
-      isWorldEntered,
-      worldChatVisible,
-      setWorldChatVisibility,
-      overlayVisible,
-      setOverlayVisibility,
-      showNames,
-      showActiveMembers,
-      shortcutUI,
-      onboarding,
-      enterXR,
-      isWebXRSupported,
-    ]
+    [worldChatVisible, overlayVisible, editorEnabled]
   );
 
   useEvent(
@@ -377,134 +142,39 @@ export function WorldView({ world }: WorldViewProps) {
     mainThread.canvas.requestPointerLock();
   }, [mainThread]);
 
-  useEffect(() => {
-    if (onboarding) document.exitPointerLock();
-  }, [onboarding]);
-
-  const onFinishOnboarding = useCallback(() => {
-    finishOnboarding();
-    mainThread.canvas?.requestPointerLock();
-  }, [mainThread.canvas, finishOnboarding]);
-
   if (isPresenting) return null;
 
   return (
     <div className="WorldView">
-      <OnboardingModal open={onboarding} world={world} requestClose={onFinishOnboarding} />
+      <WorldOnboarding world={world} />
       <Stats statsEnabled={statsEnabled} />
       <div className={classNames("WorldView__chat flex", { "WorldView__chat--open": worldChatVisible })}>
         {!("isBeingCreated" in world) && <WorldChat open={worldChatVisible} room={world} />}
       </div>
       {world && (
         <>
-          {!worldChatVisible && (
-            <Hotbar>
-              {[
-                { imageSrc: "/image/small-crate-icon.png" },
-                { imageSrc: "/image/medium-crate-icon.png" },
-                { imageSrc: "/image/large-crate-icon.png" },
-                { imageSrc: "/image/mirror-ball-icon.png" },
-                { imageSrc: "/image/black-mirror-ball-icon.png" },
-                { imageSrc: "/image/emissive-ball-icon.png" },
-              ].map((slot, index) => (
-                <HotbarSlot key={index} imageSrc={slot.imageSrc} shortcutKey={index + 1} label="Spawn Object" />
-              ))}
-            </Hotbar>
-          )}
-          <div className="WorldView__controls flex">
-            <div className="flex flex-column items-center">
-              <Tooltip content={shortcutUI ? "Hide Help" : "Show Help"}>
-                <IconButton variant="world" label="help" iconSrc={HelpIC} onClick={toggleShortcutUI} />
-              </Tooltip>
-              <Text variant="b3" color="world" weight="bold">
-                /
-              </Text>
-            </div>
-            {isWebXRSupported && (
-              <div className="flex flex-column items-center">
-                <Tooltip content="Enter XR">
-                  <IconButton variant="world" label="Enter XR" iconSrc={XRIC} onClick={enterXR} />
-                </Tooltip>
-                <Text variant="b3" color="world" weight="bold">
-                  Alt + X
-                </Text>
-              </div>
-            )}
-            <div className="flex flex-column items-center">
-              <Tooltip content={showActiveMembers ? "Hide Members" : "Show Members"}>
-                <IconButton
-                  variant="world"
-                  label="activeMembers"
-                  iconSrc={PeopleIC}
-                  onClick={toggleShowActiveMembers}
-                />
-              </Tooltip>
-              <Text variant="b3" color="world" weight="bold">
-                P
-              </Text>
-            </div>
-            <div className="flex flex-column items-center">
-              <Tooltip content={showNames ? "Hide Names" : "Show Names"}>
-                <IconButton
-                  variant="world"
-                  label="Toggle Names"
-                  iconSrc={showNames ? SubtitlesIC : SubtitlesOffIC}
-                  onClick={toggleShowNames}
-                />
-              </Tooltip>
-              <Text variant="b3" color="world" weight="bold">
-                N
-              </Text>
-            </div>
-            {activeCall && (
-              <div className="flex flex-column items-center">
-                <MuteButton
-                  showToast={showToast}
-                  activeCall={activeCall}
-                  ref={(ref) => {
-                    muteBtnRef.current = ref;
-                  }}
-                />
-                <Text variant="b3" color="world" weight="bold">
-                  M
-                </Text>
-              </div>
-            )}
-            <div className="flex flex-column items-center">
-              <Tooltip content="Disconnect">
-                <IconButton variant="danger" label="Disconnect" iconSrc={CallCrossIC} onClick={exitWorld} />
-              </Tooltip>
-              <Text variant="b3" color="world" weight="bold">
-                Alt + L
-              </Text>
-            </div>
-          </div>
+          {!worldChatVisible && <HotbarControls />}
+          <WorldControls
+            className="WorldView__controls"
+            session={session}
+            world={world}
+            activeCall={activeCall}
+            showToast={showToast}
+            isWebXRSupported={isWebXRSupported}
+            enterXR={enterXR}
+            showNames={showNames}
+            setShowNames={setShowNames}
+          />
         </>
       )}
       {world && editorEnabled && <EditorView />}
       {!("isBeingCreated" in world) && <Nametags room={world} show={showNames && !overlayVisible} />}
-      {!("isBeingCreated" in world) && (
-        <>
-          <Dialog open={showActiveMembers} onOpenChange={setShowActiveMembers}>
-            <MemberListDialog room={world} requestClose={() => setShowActiveMembers(false)} />
-          </Dialog>
-          <Dialog open={shortcutUI} onOpenChange={setShortcutUI}>
-            <Header
-              left={<HeaderTitle size="lg">Controls</HeaderTitle>}
-              right={<IconButton iconSrc={CrossIC} onClick={toggleShortcutUI} label="Close" />}
-            />
-            <div className="flex" style={{ height: "600px" }}>
-              <Scroll type="hover">
-                <ShortcutUI />
-              </Scroll>
-            </div>
-          </Dialog>
-        </>
+
+      {!overlayVisible && showNames && activeCall && (
+        <EntityTooltip session={session} world={world} activeCall={activeCall} />
       )}
-      {!overlayVisible && showNames && activeEntity && (
-        <EntityTooltip activeEntity={activeEntity} portalProcess={portalProcess} />
-      )}
-      {!overlayVisible && <Reticle activeEntity={activeEntity} mouseDown={mouseDown} />}
+      {!overlayVisible && <Reticle />}
+
       <div className="WorldView__toast-container">
         <div className={classNames("WorldView__toast", { "WorldView__toast--shown": toastShown })}>
           <Text variant="b2" color="world" weight="semi-bold">
